Using Rails in api_only mode
Api only apps won't need all controller modules used in normal applications,
for instance: ViewPaths, Layouts, Templates, and so on. Moreover, api only apps
won't need all middlewares used in normal applications, for instance:
MethodOverride (used to override the HTTP verb when passing a hidden
field in form submissions), ContentSecurityPolicy (which is a browser-only thing), and so.

One consideraion: Devise controllers will always
inherit from ApplicationController, hence we still need to use the
ActAsApiRequest concern in Devise controllers. This is because Devise is shared
between the API-only app and the admin panel.