Updating rubocop (#275)
* Updating rubocop

This PR updates the version of rubocop to 1.2 which is the latest one as
of today.

Corrections are applied as well.

* Fix rubocop offense

* Updating rubocop-rootstrap

* Removing rescue from Exception

Server errors will be logged by Rails by default. No need to rescue
from Exception to do this. The response will have a 500 HTTP code as
expected and the client shouldn't care about the body of that response.

* Removing unused translation