Add arctic_admin gem
If you attempt to run a fresh installation of this repo,
the tests won't pass because of the importations:
- As the arctic_admin gem is not initially on the gemfile,
it will cause the system specs to fail.
- We also need to have the `active_admin.js` in order to pass the specs
(and ensure functionality).

Worth mentioning that `arctic_admin/base` already adds `mixins` and `fontawesome`,
so there might be no need to add them manually.